fix: Change route parameters from i64 to i32 for database compatibility
- Fixed get_vacancy_details route parameter type
- Fixed ignore_vacancy route parameter type
- Fixed toggle_tag route parameter type
- Fixed delete_tag route parameter type
- Fixed SendMessageRequest.chat_id type

+++ server/src/jobs/models.rs
@@ -143,7 +143,7 @@ pub struct GenerateTagsRequest {
 
 #[derive(Debug, Deserialize)]
 pub struct SendMessageRequest {
-    pub chat_id: i64,
+    pub chat_id: i32,
     pub text: String,
 }
